moved review routes to under products
