+++ server/app/configure/index.js
@@ -18,7 +18,6 @@ module.exports = function (app) {
     app.use('/api/orders', require('../routes/orders'));
     app.use('/api/users', require('../routes/users'));
     app.use('/api/categories', require('../routes/categories'));
-    app.use('/api/reviews', require('../routes/reviews'));
 
 
     // Logging middleware, set as application
