+++ server/app/routes/products/index.js
@@ -3,7 +3,7 @@ var router = require('express').Router();
 
 var mongoose = require('mongoose');
 var Product = mongoose.model('Product');
-
+var Review = mongoose.model('Review');
 
 router.get('/', function(req, res, next){
 	Product.find({})
@@ -22,6 +22,43 @@ router.get('/:id', function(req, res, next){
 	})
 	.then(null, next);
 });
+
+router.delete('/:id', function(req, res, next){
+	Product.findOneAndRemove({_id: req.params.id})
+	.then(function(response){
+		res.send(response);
+	}, next);
+});
+
+
+router.get('/:id/reviews', function(req, res, next){
+	Review.find({product: req.params.id})
+	.then(function(reviews){
+		res.json(reviews);
+	}, next);
+});
+
+
+router.post('/:id/reviews', function(req, res, next){
+	var newReview = new Review({
+		user: req.body.user,
+		product: req.params.id,
+		rating: req.body.rating,
+		description: req.body.description
+	});
+	newReview.save()
+	.then(function(response){
+		res.send(response);
+	});
+});
+
+router.delete('/:id/reviews', function(req, res, next){
+	Review.findOneAndRemove({_id: req.params.id})
+	.then(function(response){
+		res.send(response);
+	}, next);
+});
+
 
 
 module.exports = router;
